Add project owners group
Rather than manually adding people we should have a group that we can
trusted people to and have that bestow the appropriate permissions (in
this case ownership of the project). A new group is used for this as
it's likely that people from many teams will need these permissions but
it shouldn't be everyone in every engineering team.