Add Dex
This sets up dex in its own namespace. Annoyingly it can't run behind
the IAP load balancer (as gitops needs to be able to hit it on the
issuerURL and I can't figure out how to have it authenticate itself with
IAP). This shouldn't be too much of a problem as only authentication
endpoints exist on it and they all refer back to the IAPd gitops.

RBAC will be configured in a later commit